Move AvatarComponent to the Beta module and remove suffix
