+++ docs/content/components/betaavatar.md
@@ -1,8 +1,8 @@
 ---
-title: Avatar
+title: BetaAvatar
 status: Beta
-source: https://github.com/primer/view_components/tree/main/app/components/primer/avatar_component.rb
-storybook: https://primer.style/view-components/stories/?path=/story/primer-avatar-component
+source: https://github.com/primer/view_components/tree/main/app/components/primer/avatar.rb
+storybook: https://primer.style/view-components/stories/?path=/story/primer-beta-avatar-component
 ---
 
 import Example from '../../src/@primer/gatsby-theme-doctocat/components/example'
