+++ docs/src/@primer/gatsby-theme-doctocat/nav.yml
@@ -17,7 +17,7 @@
     - title: AutoCompleteItem
       url: /components/betaautocompleteitem
     - title: Avatar
-      url: /components/avatar
+      url: /components/betaavatar
     - title: AvatarStack
       url: /components/avatarstack
     - title: Blankslate
